% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABR1.R
\name{AllCButAB}
\alias{AllCButAB}
\title{copies 5 columns into the new dataframe
and generates 2 new columns, 
one containing the year and one containing another grouping of the veterinary species in 5 categories.
12}
\usage{
AllCButAB(Biomic, C_Biomic, df)
}
\arguments{
\item{Biomic}{Dataframe containing old columns}

\item{C_Biomic}{minimal set of antibiotics}

\item{df}{dataframe containing the column copies}
}
\value{
df
}
\description{
copies 5 columns into the new dataframe
and generates 2 new columns, 
one containing the year and one containing another grouping of the veterinary species in 5 categories.
12
}
\examples{
 df <-AllCButAB(Biomic,C_Biomic, df)
}
\author{
Miguel Sautié PIAAS
}
