% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABR1.R
\name{an}
\alias{an}
\title{Extracts the year from the date 
8}
\usage{
an(str)
}
\arguments{
\item{str}{date string}
}
\value{
s         four-digit year
}
\description{
Extracts the year from the date 
8
}
\examples{
 y <-an(yearstring)
}
\author{
Miguel Sautié PIAAS
}
