% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABR1.R
\name{Animal}
\alias{Animal}
\title{Generates the Animal column in the newdf dataframe 
based on the grouping of the veterinary species in 5 groups: 
"ruminants", "aviaire", "horse", "swine" and "other".
11}
\usage{
Animal(Biomic, newdf)
}
\arguments{
\item{Biomic}{Dataframe containing the old ANIMAL column}

\item{newdf}{dataframe containing the new Animal column}
}
\value{
c
}
\description{
Generates the Animal column in the newdf dataframe 
based on the grouping of the veterinary species in 5 groups: 
"ruminants", "aviaire", "horse", "swine" and "other".
11
}
\examples{
 df <-Animal(Biomic,df)
}
\author{
Miguel Sautié PIAAS
}
