% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{APEC}
\alias{APEC}
\title{Boolean function to define the APEC pathotype
23}
\usage{
APEC(str)
}
\arguments{
\item{str}{string}
}
\value{
boolean value
}
\description{
Boolean function to define the APEC pathotype
23
}
\examples{
 b <-APEC(str)
}
\author{
Miguel Sautié PIAAS
}
