% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABR1.R
\name{Bmatrix}
\alias{Bmatrix}
\title{function for the conversion of bacterial colony diameters into three-valued categorical variable (S, I, R)
for a given isolate and antibiotic
7}
\usage{
Bmatrix(SABmatR, SABmatS, SBiomic)
}
\arguments{
\item{SABmatR}{resistance threshold for a given antibiotic}

\item{SABmatS}{susceptibility threshold for a given antibiotic}

\item{SBiomic}{bacterial colony diameter}
}
\value{
c         three-valued categorical variable
}
\description{
function for the conversion of bacterial colony diameters into three-valued categorical variable (S, I, R)
for a given isolate and antibiotic
7
}
\examples{
 st <-Bmatrix(TR, TS, S)
}
\author{
Miguel Sautié PIAAS
}
