% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathCountFunctions.R
\name{countcas}
\alias{countcas}
\title{case count
40}
\usage{
countcas(df, c = 1)
}
\arguments{
\item{df}{dataframe Omnilab}

\item{c}{type of count, absolute or relative.}
}
\value{
count
}
\description{
case count
40
}
\examples{
 n<-countcas(fOmniL)
}
\author{
Miguel Sautié (PIAAS)
}
