% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathCountFunctions.R
\name{countcasneg1}
\alias{countcasneg1}
\title{Negative case count. The output of this function is affected by the absence of 4007 for a given case or sample
41}
\usage{
countcasneg1(df, c = 1)
}
\arguments{
\item{df}{dataframe Omnilab}

\item{c}{type of count, absolute or relative.}
}
\value{
count
}
\description{
Negative case count. The output of this function is affected by the absence of 4007 for a given case or sample
41
}
\examples{
 nc<-countcasneg1(fOmniL)
}
\author{
Miguel Sautié (PIAAS)
}
