% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathCountFunctions.R
\name{countcasneg2}
\alias{countcasneg2}
\title{Negative case count 2.
The output of this function is affected by the absence of 4007 for a given case or sample.
44}
\usage{
countcasneg2(df)
}
\arguments{
\item{df}{dataframe Omnilab}
}
\value{
count
}
\description{
Negative case count 2.
The output of this function is affected by the absence of 4007 for a given case or sample.
44
}
\examples{
 nc<-countcasneg2(fOmniL)
}
\author{
Miguel Sautié (PIAAS)
}
