% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathCountFunctions.R
\name{countcaspos1}
\alias{countcaspos1}
\title{Positive case count. The output of this function is affected by the absence of 4007 for a given case or sample.
42}
\usage{
countcaspos1(df, c = 1)
}
\arguments{
\item{df}{dataframe Omnilab}

\item{c}{type of count, absolute or relative.}
}
\value{
count
}
\description{
Positive case count. The output of this function is affected by the absence of 4007 for a given case or sample.
42
}
\examples{
 nc<-countcaspos1(fOmniL)
}
\author{
Miguel Sautié (PIAAS)
}
