% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathCountFunctions.R
\name{countcaspos2}
\alias{countcaspos2}
\title{positive case count 2. The output of this function is affected by the absence of 4007 for a given case or sample.
43}
\usage{
countcaspos2(df)
}
\arguments{
\item{df}{dataframe Omnilab}
}
\value{
count
}
\description{
positive case count 2. The output of this function is affected by the absence of 4007 for a given case or sample.
43
}
\examples{
 nc<-countcaspos2(fOmniL)
}
\author{
Miguel Sautié (PIAAS)
}
