% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathCountFunctions.R
\name{countIsolates}
\alias{countIsolates}
\title{isolate counting
46}
\usage{
countIsolates(fOmniL2)
}
\arguments{
\item{fOmniL2}{dataframe Omnilab}
}
\value{
count
}
\description{
isolate counting
46
}
\examples{
 nc<-countIsolates(fOmniL)
}
\author{
Miguel Sautié (PIAAS)
}
