% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathCountFunctions.R
\name{countsamples}
\alias{countsamples}
\title{Sample counting.
The output of this function is affected by the absence of 4007 for a given sample.
The more samples that are not associated with any 4007.The more it underestimates the number of cases
45}
\usage{
countsamples(df, c = 1, corr = 0)
}
\arguments{
\item{df}{dataframe Omnilab}

\item{c}{type of count, absolute or relative.}

\item{corr}{correction factor}
}
\value{
count
}
\description{
Sample counting.
The output of this function is affected by the absence of 4007 for a given sample.
The more samples that are not associated with any 4007.The more it underestimates the number of cases
45
}
\examples{
 nc<-countsamples(fOmniL)
}
\author{
Miguel Sautié (PIAAS)
}
