% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{EAEC}
\alias{EAEC}
\title{Boolean function to define the EAEC pathotype
20}
\usage{
EAEC(str)
}
\arguments{
\item{str}{string}
}
\value{
boolean value
}
\description{
Boolean function to define the EAEC pathotype
20
}
\examples{
 b <-EAEC(str)
}
\author{
Miguel Sautié PIAAS
}
