% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{EIEC}
\alias{EIEC}
\title{Boolean function to define the EIEC pathotype
19}
\usage{
EIEC(str)
}
\arguments{
\item{str}{string}
}
\value{
boolean value
}
\description{
Boolean function to define the EIEC pathotype
19
}
\examples{
 b <-EIEC(str)
}
\author{
Miguel Sautié PIAAS
}
