% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{EPEC}
\alias{EPEC}
\title{Boolean function to define the EPEC pathotype
17}
\usage{
EPEC(str)
}
\arguments{
\item{str}{string}
}
\value{
boolean value
}
\description{
Boolean function to define the EPEC pathotype
17
}
\examples{
 b <-EPEC(str)
}
\author{
Miguel Sautié PIAAS
}
