% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{ETEC}
\alias{ETEC}
\title{Boolean function to define the ETEC pathotype
21}
\usage{
ETEC(str)
}
\arguments{
\item{str}{string}
}
\value{
boolean value
}
\description{
Boolean function to define the ETEC pathotype
21
}
\examples{
 b <-ETEC(str)
}
\author{
Miguel Sautié PIAAS
}
