% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{EXPEC}
\alias{EXPEC}
\title{Boolean function to define the EXPEC pathotype
18}
\usage{
EXPEC(str)
}
\arguments{
\item{str}{string}
}
\value{
boolean value
}
\description{
Boolean function to define the EXPEC pathotype
18
}
\examples{
 b <-EXPEC(str)
}
\author{
Miguel Sautié PIAAS
}
