% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fAge}
\alias{fAge}
\title{extracts age from the "Race" column
36}
\usage{
fAge(fOmniL, coname = "Race", df = newdf)
}
\arguments{
\item{fOmniL}{dataframe Omnilab}

\item{df}{new  dataframe (newdf)}
}
\value{
dataframe
}
\description{
extracts age from the "Race" column
36
}
\examples{
 df <-fAge(fOmniL), old:df <-fRace(fOmniL)
}
\author{
Miguel Sautié PIAAS
}
