% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fDescEchantillon}
\alias{fDescEchantillon}
\title{extracts sample description from the Nom.de.l.animal" column
35}
\usage{
fDescEchantillon(fOmniL, coname = "Nom.de.l.animal", df = newdf)
}
\arguments{
\item{fOmniL}{dataframe Omnilab}

\item{df}{new  dataframe (newdf)}
}
\value{
dataframe
}
\description{
extracts sample description from the Nom.de.l.animal" column
35
}
\examples{
 df <-fDescEchantillon(fOmniL)
}
\author{
Miguel Sautié PIAAS
}
