% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fHemolyse}
\alias{fHemolyse}
\title{Hémolyse
ddetermine whether a given isolat is hemolytic or not based on the "Résultat" column
33}
\usage{
fHemolyse(fOmniL, coname = "Résultat", df = newdf)
}
\arguments{
\item{fOmniL}{dataframe Omnilab}

\item{df}{new  dataframe (newdf)}
}
\value{
dataframe
}
\description{
Hémolyse
ddetermine whether a given isolat is hemolytic or not based on the "Résultat" column
33
}
\examples{
 df <-fHemolyse(fOmniL)
}
\author{
Miguel Sautié PIAAS
}
