% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fOrigin}
\alias{fOrigin}
\title{extracts Origine from the "Espèce" column
39}
\usage{
fOrigin(fOmniL, coname = "Espèce", df = newdf)
}
\arguments{
\item{fOmniL}{dataframe Omnilab}

\item{df}{new  dataframe (newdf)}
}
\value{
dataframe
}
\description{
extracts Origine from the "Espèce" column
39
}
\examples{
 df <-fOrigin(fOmniL)
}
\author{
Miguel Sautié PIAAS
}
