% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fpathotypes}
\alias{fpathotypes}
\title{determines the pathotype for a given isolat based on the "Résultat" column
28}
\usage{
fpathotypes(fOmniL, coname = "Résultat", df = newdf)
}
\arguments{
\item{fOmniL}{dataframe Omnilab}

\item{df}{dataframe}
}
\value{
dataframe
}
\description{
determines the pathotype for a given isolat based on the "Résultat" column
28
}
\examples{
 df <-fpathotypes(fOmniL)
}
\author{
Miguel Sautié PIAAS
}
