% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fPathovirotype}
\alias{fPathovirotype}
\title{Pathovirotypes
determine the pathovirotype for a given isolat based on the "Résultat" column
32}
\usage{
fPathovirotype(df = newdf)
}
\arguments{
\item{df}{new  dataframe (newdf)}

\item{fOmniL}{dataframe Omnilab}
}
\value{
dataframe
}
\description{
Pathovirotypes
determine the pathovirotype for a given isolat based on the "Résultat" column
32
}
\examples{
 df <-fPathovirotype()
}
\author{
Miguel Sautié PIAAS
}
