% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fPays}
\alias{fPays}
\title{extracts pays from the "Ville.du.propriétaire.de.l.animal" column
38}
\usage{
fPays(fOmniL, coname = "Ville.du.propriétaire.de.l.animal", df = newdf)
}
\arguments{
\item{fOmniL}{dataframe Omnilab}

\item{df}{new  dataframe (newdf)}
}
\value{
dataframe
}
\description{
extracts pays from the "Ville.du.propriétaire.de.l.animal" column
38
}
\examples{
 df <-fPays(fOmniL)
}
\author{
Miguel Sautié PIAAS
}
