% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fProvince}
\alias{fProvince}
\title{extracts province from the "Ville.du.propriétaire.de.l.animal" column
37}
\usage{
fProvince(fOmniL, coname = "Ville.du.propriétaire.de.l.animal", df = newdf)
}
\arguments{
\item{fOmniL}{dataframe Omnilab}

\item{df}{new  dataframe (newdf)}
}
\value{
dataframe
}
\description{
extracts province from the "Ville.du.propriétaire.de.l.animal" column
37
}
\examples{
 df <-fProvince(fOmniL)
}
\author{
Miguel Sautié PIAAS
}
