% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fratio_isolats}
\alias{fratio_isolats}
\title{extracts ratio_isolats from the "Résultat" column
34}
\usage{
fratio_isolats(fOmniL, coname = "Résultat", df = newdf)
}
\arguments{
\item{fOmniL}{dataframe Omnilab}

\item{df}{new  dataframe (newdf)}
}
\value{
dataframe
}
\description{
extracts ratio_isolats from the "Résultat" column
34
}
\examples{
 df <-fratio_isolats(fOmniL)
}
\author{
Miguel Sautié PIAAS
}
