% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regles3.R
\name{funM2}
\alias{funM2}
\title{search in the Biomic No_Ref column for the identifier x or its transformed version according to
the simplified M prefix rule 
1}
\usage{
funM2(x, Biomic)
}
\arguments{
\item{x}{Identifier string}

\item{Biomic}{dataframe}
}
\value{
x  the original (or transformed )identifier
}
\description{
search in the Biomic No_Ref column for the identifier x or its transformed version according to
the simplified M prefix rule 
1
}
\examples{
 st <- funM2(s, Biomic)
}
\author{
Miguel Sautié PIAAS
}
