% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regles3.R
\name{funSHYST}
\alias{funSHYST}
\title{Looks for the identifier x, or its transformed version 
in the No_Ref column of df2 (Biomic). 
he transformation rules for SHY and STF prefixes are considered.
3}
\usage{
funSHYST(
  x,
  df2,
  sts = c("SHY23", "SHY2023", "SHY-2023", "STF-2023", "STF-2022")
)
}
\arguments{
\item{x}{string identifier}

\item{df2}{dataframe}
}
\value{
s the original (or transformed )identifier
}
\description{
Looks for the identifier x, or its transformed version 
in the No_Ref column of df2 (Biomic). 
he transformation rules for SHY and STF prefixes are considered.
3
}
\examples{
 s<- funSHYST(st, df) 
 
}
\author{
Miguel Sautié PIAAS
}
