% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regles3.R
\name{funST2}
\alias{funST2}
\title{Extract the st2 suffix remaining after removing the maximum size zero prefix.
The parameter dd is defined as the number of zeros that are ignored because they are not part of 
the st2 prefix of zeros .
2}
\usage{
funST2(st2, dd)
}
\arguments{
\item{st2}{string identifier}

\item{dd}{integer}
}
\value{
the original (or transformed )identifier
}
\description{
Extract the st2 suffix remaining after removing the maximum size zero prefix.
The parameter dd is defined as the number of zeros that are ignored because they are not part of 
the st2 prefix of zeros .
2
}
\examples{
 st<-funST2(st2, 3) 
 
}
\author{
Miguel Sautié PIAAS
}
