% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fVirotypeDterminants}
\alias{fVirotypeDterminants}
\title{detects the virotype determinants for a given isolat based on the "Résultat" column
31}
\usage{
fVirotypeDterminants(fOmniL, coname = "Résultat", df = newdf, v = C_OmniDet)
}
\arguments{
\item{fOmniL}{dataframe Omnilab}

\item{df}{new  dataframe (newdf)}

\item{v}{gene set for detecting virotypes determinants (C_OmniDet)}
}
\value{
dataframe
}
\description{
detects the virotype determinants for a given isolat based on the "Résultat" column
31
}
\examples{
 df <-fVirotypeDterminants(OmniL)
}
\author{
Miguel Sautié PIAAS
}
