% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fVirotypeElargis}
\alias{fVirotypeElargis}
\title{#VirotypeElargis
determines the extended virotype for a given isolat based on the "Résultat" column
30}
\usage{
fVirotypeElargis(fOmniL, coname = "Résultat", df = newdf, v = C_OmniVE)
}
\arguments{
\item{fOmniL}{dataframe Omnilab}

\item{df}{new  dataframe (newdf)}

\item{v}{gene set for determining extended virotypes (C_OmniVE)}
}
\value{
dataframe
}
\description{
#VirotypeElargis
determines the extended virotype for a given isolat based on the "Résultat" column
30
}
\examples{
 df <-fVirotypeElargis(fOmniL)
}
\author{
Miguel Sautié PIAAS
}
