% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{fVirotypeStandard}
\alias{fVirotypeStandard}
\title{determines the standard virotype for a given isolat based on the "Résultat" column
29}
\usage{
fVirotypeStandard(fOmniL, coname = "Résultat", df = newdf, v = C_Omni)
}
\arguments{
\item{fOmniL}{dataframe Omnilab}

\item{df}{dataframe}

\item{C_Omni}{gene set for determining virotypes}
}
\value{
dataframe
}
\description{
determines the standard virotype for a given isolat based on the "Résultat" column
29
}
\examples{
 df <-fVirotypeStandard(fOmniL)
}
\author{
Miguel Sautié PIAAS
}
