% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{hrAPEC}
\alias{hrAPEC}
\title{Boolean function to define the hrAPEC pathotype
25}
\usage{
hrAPEC(str)
}
\arguments{
\item{str}{string}
}
\value{
boolean value
}
\description{
Boolean function to define the hrAPEC pathotype
25
}
\examples{
 b <-hrAPEC(str)
}
\author{
Miguel Sautié PIAAS
}
