% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{joinNames}
\alias{joinNames}
\title{Concatenates strings
26}
\usage{
joinNames(st0, st, sep = ":")
}
\arguments{
\item{st0}{string 1}

\item{st}{string 2}

\item{sep}{separator(dv,:)}
}
\value{
boolean value
}
\description{
Concatenates strings
26
}
\examples{
 b <-joinNames(str, str2)
}
\author{
Miguel Sautié PIAAS
}
