% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{mvPathoTypes}
\alias{mvPathoTypes}
\title{missing value for pathotypes (Other)
27}
\usage{
mvPathoTypes(st0, patho, mv = NA)
}
\arguments{
\item{st0}{string 1}

\item{patho}{string 2}

\item{mv}{missing value}
}
\value{
boolean value
}
\description{
missing value for pathotypes (Other)
27
}
\examples{
 b <-mvPathoTypes(str, patho)
}
\author{
Miguel Sautié PIAAS
}
