% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{NTEC2}
\alias{NTEC2}
\title{Boolean function to define the NTEC2 pathotype
16}
\usage{
NTEC2(str)
}
\arguments{
\item{str}{string}
}
\value{
boolean value
}
\description{
Boolean function to define the NTEC2 pathotype
16
}
\examples{
 b <-NTEC2(str)
}
\author{
Miguel Sautié PIAAS
}
