% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{probAPEC}
\alias{probAPEC}
\title{Boolean function to define the probAPEC pathotype
24}
\usage{
probAPEC(str)
}
\arguments{
\item{str}{string}
}
\value{
boolean value
}
\description{
Boolean function to define the probAPEC pathotype
24
}
\examples{
 b <-probAPEC(str)
}
\author{
Miguel Sautié PIAAS
}
