% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABR1.R
\name{profileAMR}
\alias{profileAMR}
\title{Generates the profileAMR column in the newdf dataframe 
from the ternary matrix (I, R, S) (Biomic) 
and the minimum set of common antibiotics (C_Biomic).
10}
\usage{
profileAMR(Biomic, C_Biomic, newdf)
}
\arguments{
\item{Biomic}{Dataframe containing ternary matrix (I, R, S) for a given set of isolates and antibiotics}

\item{C_Biomic}{The minimum set of common antibiotics}

\item{newdf}{}
}
\value{
c
}
\description{
Generates the profileAMR column in the newdf dataframe 
from the ternary matrix (I, R, S) (Biomic) 
and the minimum set of common antibiotics (C_Biomic).
10
}
\examples{
 st <-profileAMR(Biomic, CBiomic, st)
}
\author{
Miguel Sautié PIAAS
}
