% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABR1.R
\name{profileMDR1}
\alias{profileMDR1}
\title{Generates a new column (profileMDR) in the dataframe df
13}
\usage{
profileMDR1(newdf, C_Family, C_Biomic, Lmdr = 1)
}
\arguments{
\item{newdf}{new generated dataframe}

\item{C_Family}{families to which the antibiotics in the minimum set of common antibiotics belong}

\item{C_Biomic}{minimal set of common antibiotics}

\item{Lmdr}{minimum number of antibiotic families (Lmdr+2) to which a given isolate is resistant to define the value "MDR" or "Non MDR".}
}
\value{
newdf
}
\description{
Generates a new column (profileMDR) in the dataframe df
13
}
\examples{
 df <-profileMDR1(Biomic,C_Biomic, df)
}
\author{
Miguel Sautié PIAAS
}
