% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathCountFunctions.R
\name{samplecount2}
\alias{samplecount2}
\title{Sample counting 2
The output of this function is not affected by the absence of 4007 for a given sample.
53}
\usage{
samplecount2(d)
}
\arguments{
\item{df}{dataframe Omnilab}
}
\value{
count
}
\description{
Sample counting 2
The output of this function is not affected by the absence of 4007 for a given sample.
53
}
\examples{
 nc<-samplecount2(fOmniL)
}
\author{
Miguel Sautié (PIAAS)
}
