% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regles3.R
\name{selConversion}
\alias{selConversion}
\title{Search in Biomic for OmniL identifiers or their modified versions
according to the 3 transformation rules for prefixes M, SHY and STF.
The output: OmniL with the original identifiers or the modified versions of them 
found in Biomic as well as the Biomic  dataframe with additional lines corresponding 
to the non-prefixed identifiers in the OmniLab file. 
This function is intended to be run before merging the two dataframes 
with the native merge function. 
4}
\usage{
selConversion(OmniL, Biomic)
}
\arguments{
\item{Biomic}{Biomic dataframe}

\item{Omnil}{Omnilab dataframe}
}
\value{
list of both modified dataframes
}
\description{
Search in Biomic for OmniL identifiers or their modified versions
according to the 3 transformation rules for prefixes M, SHY and STF.
The output: OmniL with the original identifiers or the modified versions of them 
found in Biomic as well as the Biomic  dataframe with additional lines corresponding 
to the non-prefixed identifiers in the OmniLab file. 
This function is intended to be run before merging the two dataframes 
with the native merge function. 
4
}
\examples{
 Ldf<-selConversion(df1, df2) 
}
\author{
Miguel Sautié PIAAS
}
