% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regles3.R
\name{SselConversion}
\alias{SselConversion}
\title{simplified version of selConversion that does not add non-prefixed identifiers to Biomic
5}
\usage{
SselConversion(OmniL, Biomic)
}
\arguments{
\item{Biomic}{Biomic dataframe}

\item{Omnil}{Omnilab dataframe}
}
\value{
modified OmniLab dataframe
}
\description{
simplified version of selConversion that does not add non-prefixed identifiers to Biomic
5
}
\examples{
 mdf-SselConversion(df1, df2) 
}
\author{
Miguel Sautié PIAAS
}
