% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{STEC}
\alias{STEC}
\title{Boolean function to define the STEC pathotype
22}
\usage{
STEC(str)
}
\arguments{
\item{str}{string}
}
\value{
boolean value
}
\description{
Boolean function to define the STEC pathotype
22
}
\examples{
 b <-STEC(str)
}
\author{
Miguel Sautié PIAAS
}
