% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABR1.R
\name{Transformation3States}
\alias{Transformation3States}
\title{generates categorical values (R, S or I) from colony diameters per isolate and per antibiotic  (Biomic),
from a minimum set of common antibiotics (C_Biomic) and a set of resistance and susceptibility thresholds
(ABmat).  
9}
\usage{
Transformation3States(Biomic, C_Biomic, ABmat, newdf)
}
\arguments{
\item{Biomic}{dataframe containing colony diameters per isolate and per antibiotic}

\item{C_Biomic}{minimum set of common antibiotics}

\item{ABmat}{dataframe containning the set of resistance and susceptibility thresholds}

\item{newdf}{the new generated dataframe}
}
\value{
newdf
}
\description{
generates categorical values (R, S or I) from colony diameters per isolate and per antibiotic  (Biomic),
from a minimum set of common antibiotics (C_Biomic) and a set of resistance and susceptibility thresholds
(ABmat).  
9
}
\examples{
 df<-Transformation3States(Biomic,C_Biomic,ABmat,df)
 
}
\author{
Miguel Sautié PIAAS
}
