% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathFunctions.R
\name{year}
\alias{year}
\title{Extracts year from the date
15}
\usage{
year(str, sep = "-")
}
\arguments{
\item{str}{date string}
}
\value{
newdf
}
\description{
Extracts year from the date
15
}
\examples{
 df <-year(strDate)
}
\author{
Miguel Sautié  PIAAS
}
